Fix Death Particle

+++ src/main/java/com/test/eraser/additional/ModTiers.java
@@ -1,6 +1,7 @@
 package com.test.eraser.additional;
 
 import net.minecraft.tags.BlockTags;
+import net.minecraft.tags.InstrumentTags;
 import net.minecraft.world.item.Tier;
 import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraftforge.common.ForgeTier;
@@ -10,7 +11,7 @@ public class ModTiers {
             1919, // harvestLevel
             0, // durability
             50.0F, // miningSpeed
-            Float.MAX_VALUE, // attackDamageBonus
+            Integer.MAX_VALUE, // attackDamageBonus
             30, // enchantmentValue
             BlockTags.NEEDS_DIAMOND_TOOL,
             () -> Ingredient.EMPTY
@@ -19,7 +20,7 @@ public class ModTiers {
             1919,
             0,
             15.0F,
-            Float.MAX_VALUE,
+            Integer.MAX_VALUE,
             30,
             BlockTags.NEEDS_DIAMOND_TOOL,
             () -> Ingredient.EMPTY
